MVP-6450: init AuthManager instance
