+++ packages/notifi-frontend-client/lib/client/index.ts
@@ -2,3 +2,4 @@ export * from './NotifiFrontendClient';
 export * from './NotifiSmartLinkClient';
 export { instantiateFrontendClient, newSmartLinkClient } from './clientFactory';
 export * from './deprecated';
+export * from './AuthManager';
